+++ src/api/index.ts
@@ -1,3 +1,4 @@
+export * from "./cds"
 export * from "./delete"
 export * from "./discovery"
 export * from "./nodeContents"