+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+- feature check
+
 ## [0.5.35] 2019-12-12
 
 ### Fixed