+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Format based on [Keep a Changelog](http://keepachangelog.com/)
 - syntax check for CDS views
 - CDS annotations definition(raw)
 - DDIC components (required for CDS completions)
+- DDIC elements
 
 ## [0.5.35] 2019-12-12